Extract deserialize of monitoring frame msg
