+++ include/psen_scan_v2/monitoring_frame_msg.h
@@ -23,8 +23,6 @@
 #include <string>
 #include <vector>
 
-#include <gtest/gtest_prod.h>
-
 #include "psen_scan_v2/raw_scanner_data.h"
 #include "psen_scan_v2/tenth_of_degree.h"
 
@@ -38,50 +36,8 @@ static constexpr uint32_t MAX_SCANNER_ID{ 0x03 };
 static constexpr uint16_t NUMBER_OF_BYTES_SCAN_COUNTER{ 4 };
 static constexpr uint16_t NUMBER_OF_BYTES_SINGLE_MEASURE{ 2 };
 
-class FieldHeader
-{
-public:
-  using Id = uint8_t;
-  using Length = uint16_t;
-
-public:
-  FieldHeader(Id id, Length length);
-
-public:
-  Id id() const;
-  Length length() const;
-
-  static std::string idToString(Id id);
-
-private:
-  Id id_;
-  Length length_;
-};
-
-struct AdditionalFieldIds
-{
-  static constexpr FieldHeader::Id SCAN_COUNTER{ 0x02 };
-  static constexpr FieldHeader::Id MEASURES{ 0x05 };
-  static constexpr FieldHeader::Id END_OF_FRAME{ 0x09 };
-};
-
 class MonitoringFrameMsg
 {
-public:
-  class MonitoringFrameFormatError : public std::runtime_error
-  {
-  public:
-    MonitoringFrameFormatError(const std::string& msg = "Error while decoding laser scanner measurement data");
-  };
-
-  class MonitoringFrameFormatErrorScanCounterUnexpectedSize : public MonitoringFrameFormatError
-  {
-  public:
-    MonitoringFrameFormatErrorScanCounterUnexpectedSize(const std::string& msg) : MonitoringFrameFormatError(msg)
-    {
-    }
-  };
-
 public:
   MonitoringFrameMsg() = default;
   MonitoringFrameMsg(const TenthOfDegree& from_theta,
@@ -94,7 +50,6 @@ public:
     , measures_(measures){
 
     };
-  static MonitoringFrameMsg deserialize(const MaxSizeRawData& data, const std::size_t& num_bytes);
 
 public:
   TenthOfDegree fromTheta() const;
@@ -133,23 +88,10 @@ public:
   }
 
 private:
-  using FieldId = FieldHeader::Id;
-  using FieldLength = FieldHeader::Length;
+  // FRIEND_TEST(FieldHeaderTest, testReadSuccess);
+  // FRIEND_TEST(FieldHeaderTest, testReadHeaderTooShortFailure);
 
-private:
-  static void readAngle(std::istringstream& is, double& angle);
-
-  static FieldHeader readFieldHeader(std::istringstream& is, const std::size_t& max_num_bytes);
-  static void readScanCounter(std::istringstream& is, uint32_t& scan_counter, const FieldLength length);
-  static void readMeasures(std::istringstream& is, std::vector<double>& measures, const FieldLength length);
-
-  void checkFixedFields();
-
-private:
-  FRIEND_TEST(FieldHeaderTest, testReadSuccess);
-  FRIEND_TEST(FieldHeaderTest, testReadHeaderTooShortFailure);
-
-private:
+protected:
   uint32_t device_status_fixed_{ 0 };
   uint32_t op_code_fixed_{ OP_CODE_MONITORING_FRAME };
   uint32_t working_mode_fixed_{ 0 };
@@ -160,18 +102,12 @@ private:
 
   uint32_t scan_counter_{ 0 };
   std::vector<double> measures_;
+
+public:
   friend DynamicSizeRawData serialize(MonitoringFrameMsg& frame);
+  friend MonitoringFrameMsg deserialize(const MaxSizeRawData& data, const std::size_t& num_bytes);
+  friend void checkFixedFields(MonitoringFrameMsg& msg);
 };
-
-inline FieldHeader::Id FieldHeader::id() const
-{
-  return id_;
-}
-
-inline FieldHeader::Length FieldHeader::length() const
-{
-  return length_;
-}
 
 inline TenthOfDegree MonitoringFrameMsg::fromTheta() const
 {
@@ -192,12 +128,6 @@ inline std::vector<double> MonitoringFrameMsg::measures() const
 {
   return measures_;
 }
-
-inline MonitoringFrameMsg::MonitoringFrameFormatError::MonitoringFrameFormatError(const std::string& msg)
-  : std::runtime_error(msg)
-{
-}
-
 }  // namespace psen_scan_v2
 
 #endif  // PSEN_SCAN_V2_MONITORING_FRAME_MSG_H
