+++ include/psen_scan_v2/raw_processing.h
@@ -18,6 +18,7 @@
 #include <sstream>
 #include <functional>
 #include <cmath>
+#include <algorithm>
 
 #include <fmt/core.h>
