+++ src/monitoring_frame_msg.cpp
@@ -29,110 +29,4 @@
 
 namespace psen_scan_v2
 {
-constexpr FieldHeader::Id AdditionalFieldIds::SCAN_COUNTER;
-constexpr FieldHeader::Id AdditionalFieldIds::MEASURES;
-constexpr FieldHeader::Id AdditionalFieldIds::END_OF_FRAME;
-
-FieldHeader::FieldHeader(Id id, Length length) : id_(id), length_(length)
-{
-}
-
-MonitoringFrameMsg MonitoringFrameMsg::deserialize(const MaxSizeRawData& data, const std::size_t& num_bytes)
-{
-  MonitoringFrameMsg msg;
-
-  MaxSizeRawData tmp_data{ data };
-  std::istringstream is(std::string(tmp_data.data(), tmp_data.size()));
-
-  raw_processing::read(is, msg.device_status_fixed_);
-  raw_processing::read(is, msg.op_code_fixed_);
-  raw_processing::read(is, msg.working_mode_fixed_);
-  raw_processing::read(is, msg.transaction_type_fixed_);
-  raw_processing::read(is, msg.scanner_id_fixed_);
-
-  raw_processing::read<uint16_t, TenthOfDegree>(is, msg.from_theta_fixed_);
-  raw_processing::read<uint16_t, TenthOfDegree>(is, msg.resolution_fixed_);
-
-  msg.checkFixedFields();
-
-  bool end_of_frame{ false };
-  while (!end_of_frame)
-  {
-    const FieldHeader header{ readFieldHeader(is, num_bytes) };
-
-    switch (header.id())
-    {
-      case AdditionalFieldIds::SCAN_COUNTER:
-        if (header.length() != NUMBER_OF_BYTES_SCAN_COUNTER)
-        {
-          throw MonitoringFrameFormatErrorScanCounterUnexpectedSize(fmt::format(
-              "Length of scan counter field is {}, but should be {}.", header.length(), NUMBER_OF_BYTES_SCAN_COUNTER));
-        }
-        raw_processing::read(is, msg.scan_counter_);
-        break;
-
-      case AdditionalFieldIds::MEASURES:
-        raw_processing::readArray<uint16_t, double>(is,
-                                                    msg.measures_,
-                                                    header.length() / NUMBER_OF_BYTES_SINGLE_MEASURE,
-                                                    [](uint16_t raw_element) { return raw_element / 1000.; });
-        break;
-
-      case AdditionalFieldIds::END_OF_FRAME:
-        end_of_frame = true;
-        break;
-
-      default:
-        throw MonitoringFrameFormatError(
-            fmt::format("Header Id {:#04x} unknown. Cannot read additional field of monitoring frame.", header.id()));
-    }
-  }
-
-  return msg;
-}
-
-FieldHeader MonitoringFrameMsg::readFieldHeader(std::istringstream& is, const std::size_t& max_num_bytes)
-{
-  FieldId id;
-  FieldLength length;
-  raw_processing::read(is, id);
-  raw_processing::read(is, length);
-
-  if (length >= max_num_bytes)
-  {
-    throw MonitoringFrameFormatError(
-        fmt::format("Length given in header of additional field is too large: {}, id: {:#04x}", length, id));
-  }
-  if (length > 0)
-  {
-    length--;
-  }
-  return FieldHeader(id, length);
-}
-
-void MonitoringFrameMsg::checkFixedFields()
-{
-  if (OP_CODE_MONITORING_FRAME != op_code_fixed_)
-  {
-    // TODO: Get rid of the issue not to spam the system with this debug messages
-    //       Would something like  ROS_DEBUG_THROTTLE(period, ...) be a good solution?
-    PSENSCAN_DEBUG("MonitoringFrameMsg", "Wrong Op Code!");
-  }
-
-  if (ONLINE_WORKING_MODE != working_mode_fixed_)
-  {
-    PSENSCAN_DEBUG("MonitoringFrameMsg", "Invalid working mode!");
-  }
-
-  if (GUI_MONITORING_TRANSACTION != transaction_type_fixed_)
-  {
-    PSENSCAN_DEBUG("MonitoringFrameMsg", "Invalid transaction type!");
-  }
-
-  if (MAX_SCANNER_ID < scanner_id_fixed_)
-  {
-    PSENSCAN_DEBUG("MonitoringFrameMsg", "Invalid Scanner id!");
-  }
-}
-
 }  // namespace psen_scan_v2
